Move treemap.rs from libextra to libstd
