+++ src/libextra/lib.rs
@@ -69,7 +69,6 @@ pub mod smallintmap;
 pub mod sort;
 
 pub mod dlist;
-pub mod treemap;
 pub mod btree;
 
 // And ... other stuff
