+++ src/libextra/test.rs
@@ -26,8 +26,8 @@ use stats::Stats;
 use stats;
 use term;
 use time::precise_time_ns;
-use treemap::TreeMap;
 
+use std::treemap::TreeMap;
 use std::clone::Clone;
 use std::comm::{stream, SharedChan, GenericPort, GenericChan};
 use std::io;
