+++ src/librustpkg/tests.rs
@@ -17,14 +17,14 @@ use std::io::fs;
 use std::io::File;
 use std::io::process;
 use std::io::process::ProcessExit;
+use std::treemap::TreeMap;
+use std::run::ProcessOutput;
 use extra::arc::Arc;
 use extra::arc::RWArc;
 use extra::tempfile::TempDir;
 use extra::workcache;
 use extra::workcache::{Database, Logger};
-use extra::treemap::TreeMap;
 use extra::getopts::groups::getopts;
-use std::run::ProcessOutput;
 use installed_packages::list_installed_packages;
 use package_id::{PkgId};
 use version::{ExactRevision, NoVersion, Version, Tagged};
