+++ src/libstd/os.rs
@@ -1442,7 +1442,7 @@ mod tests {
         }
     }
 
-    #[test]
+    #[test] #[ignore]
     fn memory_map_file() {
         use result::{Ok, Err};
         use os::*;
