+++ src/test/bench/core-map.rs
@@ -11,7 +11,7 @@
 extern mod extra;
 
 use extra::time;
-use extra::treemap::TreeMap;
+use std::treemap::TreeMap;
 use std::hashmap::{HashMap, HashSet};
 use std::os;
 use std::rand::{Rng, IsaacRng, SeedableRng};
